Separate out app and api
